refactor(workspace): inline runExtensionFactories into both call sites
Delete the shared run-extension-factories.ts and inline its logic
directly into withExtension() (single-entry, no loop) and
open() (multi-entry loop). The shared abstraction hid an asymmetry
between the two callers without providing meaningful deduplication.

Preserves exact error handling semantics: async fire-and-forget,
LIFO cleanup, error collection with console.error logging.
